+++ RayTracer/Tracer.cpp
@@ -1,6 +1,6 @@
 #include "Tracer.h"
 
-glm::ivec3 Tracer::traceRay(Ray _myRay)
+glm::vec3 Tracer::traceRay(Ray _myRay)
 {
 	intersectionReturn myResult;
 	Geometry geometry;
@@ -11,14 +11,14 @@ glm::ivec3 Tracer::traceRay(Ray _myRay)
 
 		if (myResult.intersection == true)
 		{
-			glm::ivec3 colour = objects[0].shadePixel(_myRay, myResult.intersectionPoint);
+			glm::vec3 colour = objects[0].shadePixel(_myRay, myResult.intersectionPoint);
 			return colour;
 			//return glm::ivec3(255, 0, 0);
 		}
 		else
 		{
 			//Background colour
-			return glm::ivec3(0, 0, 0);
+			return glm::vec3(0, 0, 0);
 		}
 	}
 }
